requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.15
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
Pillow==8.3.2
PyJWT == 2.4.0
django-enumchoicefield == 3.0.1